rename request response formatter tests
