--- content/404.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>404</title>
-</head>
-<body>
-    <div><h1>OOPSIES!<br>It's a 404!</h1></div>
-</body>
-</html>
