--- src/main/java/App.java
@@ -1,13 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-
-
-public class App {
-    public static void main(String[] args) {
-//        HelloWorldServer server = new HelloWorldServer();
-//        server.go();
-//        HelloWorldClient testClient = new HelloWorldClient();
-//        testClient.go();
-    }
-}
